+++ CHANGES.md
@@ -5,9 +5,13 @@
 * Runtime/wasm: faster copy between a Wasm string and an JavaScript array buffer (#2124)
 * Runtime/wasm: faster conversion of small JavaScript strings to Wasm (#2124)
 * Compiler: improved shape computation (#2198)
+* Add the --build-config and --apply-build-config flags (#2177)
 
 ## Bug fixes
 * Compiler/wasm: fix int division return type to Unnormalized (#2197)
+
+
+
 
 # 6.3.2 (2026-02-15) - Lille